% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cox_res.R
\name{add_sig_line_to_plot}
\alias{add_sig_line_to_plot}
\title{Add a Cox regression statistical significance line}
\usage{
add_sig_line_to_plot(cur_plot, coord_flip, add_sig_line)
}
\arguments{
\item{cur_plot}{ggplot2 object to add on to.}

\item{coord_flip}{Boolean to indicate if the ggplot2 has been flipped or not.}

\item{add_sig_line}{Boolean to indicate whether the significance line should
be added or not.}
}
\description{
This function will add a Cox regression statistical significance line to the
input ggplot2.
}
