% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_rates.R
\name{CompareKMRates}
\alias{CompareKMRates}
\title{Compare Rates}
\usage{
CompareKMRates(
  data,
  alpha = 0.05,
  arm_name = "arm",
  return_surv = TRUE,
  status_name = "status",
  tau = NULL,
  time_name = "time"
)
}
\arguments{
\item{data}{Data.frame.}

\item{alpha}{Type I error.}

\item{arm_name}{Name of arm column.}

\item{return_surv}{Logical, TRUE for survival, FALSE for cumulative incidence.}

\item{status_name}{Name of status column.}

\item{tau}{Truncation time.}

\item{time_name}{Name of time column.}
}
\value{
Data.frame.
}
\description{
Compare the Kaplan-Meier survival or incidence rates of two treatment arms.
}
