% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_cox.R
\name{CompreCoxCstat}
\alias{CompreCoxCstat}
\title{Compare Cox Models}
\usage{
CompreCoxCstat(status, time, x1, x2, folds = 10, simple_cstat = TRUE)
}
\arguments{
\item{status}{Event status.}

\item{time}{Event time.}

\item{x1}{Model matrix for the first model.}

\item{x2}{Model matrix for the second model.}

\item{folds}{Cross validation folds.}

\item{simple_cstat}{If TRUE uses the unweighted C-stat, otherwise
uses the inverse probability of censoring weighted C-stat.}
}
\description{
Compares Cox models based on two sets of predictors with respect to
their C-statistics via K-fold cross validation.
}
