% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{show,SurvSummary-method}
\alias{show,SurvSummary-method}
\title{Show Method for SurvSummary Object}
\usage{
\S4method{show}{SurvSummary}(object)
}
\arguments{
\item{object}{An object of class \code{SurvSummary}.}
}
\description{
Show Method for SurvSummary Object
}
