% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_frame.R
\name{Get2ArmParaPlotFrame}
\alias{Get2ArmParaPlotFrame}
\title{Plotting Frame for Parametric Models}
\usage{
Get2ArmParaPlotFrame(para_model, tau, eval_points = 1000, return_surv = TRUE)
}
\arguments{
\item{para_model}{Fitted parametric model.}

\item{tau}{Upper limit of observation period.}

\item{eval_points}{Number of points at which to evaluate the curve.}

\item{return_surv}{Logical, TRUE for survival, FALSE for cumulative incidence.}
}
\value{
Data.frame containing \code{time}, \code{surv}, \code{arm}.
}
\description{
Plotting Frame for Parametric Models
}
