% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_frame.R
\name{Get2ArmPlotFrame}
\alias{Get2ArmPlotFrame}
\title{Get 2 Arm Plotting Frame}
\usage{
Get2ArmPlotFrame(
  data,
  tau,
  arm_name = "arm",
  eval_points = 1000,
  return_surv = TRUE,
  status_name = "status",
  time_name = "time"
)
}
\arguments{
\item{data}{Data.frame.}

\item{tau}{Trunction time.}

\item{arm_name}{Name of arm column.}

\item{eval_points}{Number of points at which to evaluate the curve.}

\item{return_surv}{Logical, TRUE for survival, FALSE for cumulative incidence.}

\item{status_name}{Name of status column.}

\item{time_name}{Name of time column.}
}
\value{
Data.frame.
}
\description{
Get 2 Arm Plotting Frame
}
