% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_cox_summary.R
\name{get_cox_summary}
\alias{get_cox_summary}
\title{Summarizes the Cox Regression Analsysis}
\usage{
get_cox_summary(cox.res)
}
\arguments{
\item{cox.res}{A cox regression analysis results object.}
}
\value{
A data.frame that summarizes the results.
}
\description{
This function summarizes the results of a cox regression returning the
results in a data.frame. It will calculate the confidence interval.
}
\examples{
cox.res <- survival::coxph(survival::Surv(time,status) ~ rx, survival::colon)
get_cox_summary(cox.res)
}

