% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kaplan_meier.R
\name{GetCurves}
\alias{GetCurves}
\title{Generate Survival and Hazard Curves}
\usage{
GetCurves(data, alpha = 0.05, status_name = "status", time_name = "time")
}
\arguments{
\item{data}{Data.frame.}

\item{alpha}{Type I error.}

\item{status_name}{Name of status column.}

\item{time_name}{Name of time column.}
}
\value{
stepfun.
}
\description{
Intended for data from a single sample.
}
