% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_kaplan_meier.R
\name{GetKMCurve}
\alias{GetKMCurve}
\title{Get Kaplan-Meier Curve}
\usage{
GetKMCurve(
  data,
  return_surv = TRUE,
  status_name = "status",
  time_name = "time"
)
}
\arguments{
\item{data}{Data.frame.}

\item{return_surv}{Logical, TRUE for survival, FALSE for cumulative incidence.}

\item{status_name}{Name of status column.}

\item{time_name}{Name of time column.}
}
\value{
Step function.
}
\description{
Return a function that calculates the survival probability for a single
treatment arm.
}
