% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_nars.R
\name{GetKMNARs}
\alias{GetKMNARs}
\title{Get Numbers at Risk}
\usage{
GetKMNARs(
  data,
  x_breaks,
  arm_name = "arm",
  status_name = "status",
  time_name = "time"
)
}
\arguments{
\item{data}{Data.frame.}

\item{x_breaks}{Time points at which to determine the NARs.}

\item{arm_name}{Name of arm column.}

\item{status_name}{Name of status column.}

\item{time_name}{Name of time column.}
}
\value{
Data.frame containing \code{time}, \code{nar_ctrl}, \code{nar_trt}.
}
\description{
Numbers at risk for competing risks data.
}
