% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_frame.R
\name{GetKMPlotFrame}
\alias{GetKMPlotFrame}
\title{Prepare Kaplan-Meier Plotting Frame}
\usage{
GetKMPlotFrame(
  data,
  eval_points = 1000,
  return_surv = TRUE,
  status_name = "status",
  tau = NULL,
  time_name = "time"
)
}
\arguments{
\item{data}{Data.frame}

\item{eval_points}{Number of points at which to evaluate the curve.}

\item{return_surv}{Logical, TRUE for survival, FALSE for cumulative incidence.}

\item{status_name}{Name of status column.}

\item{tau}{Trunction time.}

\item{time_name}{Name of time column.}
}
\value{
Data.frame.
}
\description{
Prepare Kaplan-Meier Plotting Frame
}
