% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_rates.R
\name{GetKMSECurve}
\alias{GetKMSECurve}
\title{Get Kaplan-Meier Standard Error Curve}
\usage{
GetKMSECurve(
  data,
  return_surv = TRUE,
  status_name = "status",
  time_name = "time"
)
}
\arguments{
\item{data}{Data.frame.}

\item{return_surv}{Logical, TRUE for survival, FALSE for cumulative incidence.}

\item{status_name}{Name of status column.}

\item{time_name}{Name of time column.}
}
\value{
Step function.
}
\description{
Return a function that calculates the standard error of the survival probability
for a single treatment arm.
}
