% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nars.R
\name{GetNARCurve}
\alias{GetNARCurve}
\title{Get Number at Risk Curve}
\usage{
GetNARCurve(data, status_name = "status", time_name = "time")
}
\arguments{
\item{data}{Data.frame.}

\item{status_name}{Name of status column.}

\item{time_name}{Name of time column.}
}
\value{
Step function.
}
\description{
Return a function that calculates the number at risk for a single treatment
arm.
}
