% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{InfluenceCIC}
\alias{InfluenceCIC}
\title{CIC Influence Function}
\usage{
InfluenceCIC(status, time, trunc_time)
}
\arguments{
\item{status}{Status, coded as 0 for censoring, 1 for an event, 2 for death.}

\item{time}{Observation time.}

\item{trunc_time}{Time at which to evaluate the influence function.}
}
\value{
Tabulate cumulative incidence curve.
}
\description{
Influence function for the cumulative incidence of the status == 1 event.
}
