% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{InfluenceKM}
\alias{InfluenceKM}
\title{Influence Function R}
\usage{
InfluenceKM(status, time, trunc_time)
}
\arguments{
\item{status}{Status, coded as 0 for censoring, 1 for death.}

\item{time}{Observation time.}

\item{trunc_time}{Truncation time.}
}
\value{
Data.frame.
}
\description{
Influence function of the Kaplan-Meier estimator at time t. Specifically,
\eqn{\psi_{i}(t) = -S(t)\int_{0}^{t} dM_{i}(u) / Y(u)}.
}
