% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{InfluenceRMST}
\alias{InfluenceRMST}
\title{RMST Influence Function}
\usage{
InfluenceRMST(status, time, trunc_time)
}
\arguments{
\item{status}{Status, coded as 0 for censoring, 1 for death.}

\item{time}{Observation time.}

\item{trunc_time}{Truncation time.}
}
\value{
Numeric vector of influence function values for each observation.
}
\description{
Influence function of the restricted mean survival time at time t. Specifically,
\eqn{\psi_{i}(t) = -S(t)\int_{0}^{t} \mu_{\tau} dM_{i}(u) / Y(u)}.
}
