% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{OneSample-class}
\alias{OneSample-class}
\title{One Sample Survival Summary Object.}
\description{
One Sample Survival Summary Object.
}
\section{Slots}{

\describe{
\item{\code{CumHaz}}{Cumulative hazard curve.}

\item{\code{CumHazVar}}{Variance of cumulative hazard.}

\item{\code{CumHazLower}}{Lower bound of cumulative hazard.}

\item{\code{CumHazUpper}}{Upper bound of cumulative hazard.}

\item{\code{NAR}}{Number at risk curve.}

\item{\code{Surv}}{Survival curve.}

\item{\code{SurvVar}}{Variance of survival curve.}

\item{\code{SurvLower}}{Lower bound of survival curve.}

\item{\code{SurvUpper}}{Upper bound of survival curve.}

\item{\code{tmax}}{Max observation time.}
}}

