% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{show,OneSample-method}
\alias{show,OneSample-method}
\title{Show Method for OneSample Object}
\usage{
\S4method{show}{OneSample}(object)
}
\arguments{
\item{object}{An object of class \code{OneSample}.}
}
\description{
Show Method for OneSample Object
}
