% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{OneSampleCIC-class}
\alias{OneSampleCIC-class}
\title{One Sample Cumulative Incidence Summary Object}
\description{
One Sample Cumulative Incidence Summary Object
}
\section{Slots}{

\describe{
\item{\code{CIC}}{Cumulative incidence curve.}

\item{\code{CICVar}}{Variance of cumulative incidence curve.}

\item{\code{CICLower}}{Lower bound of cumulative incidence curve.}

\item{\code{CICUpper}}{Upper bound of cumulative incidence curve.}

\item{\code{NAR}}{Number at risk curve.}

\item{\code{tmax}}{Max observation time.}
}}

