% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{show,OneSampleCIC-method}
\alias{show,OneSampleCIC-method}
\title{Show Method for OneSampleCIC Object}
\usage{
\S4method{show}{OneSampleCIC}(object)
}
\arguments{
\item{object}{An object of class \code{OneSampleCIC}.}
}
\description{
Show Method for OneSampleCIC Object
}
