% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{show,OneSampleSurv-method}
\alias{show,OneSampleSurv-method}
\title{Show Method for OneSampleSurv Object}
\usage{
\S4method{show}{OneSampleSurv}(object)
}
\arguments{
\item{object}{An object of class \code{OneSampleSurv}.}
}
\description{
Show Method for OneSampleSurv Object
}
