% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cox_res.R
\name{plot_cox_res}
\alias{plot_cox_res}
\title{Plot Cox Regression Results}
\usage{
plot_cox_res(cox.res.df, group, xlab, ylab)
}
\arguments{
\item{group}{Vector containing the column name(s) of the grouping variables.
This is only applicable if multivariate cox regression was run.
The data will be faceted for each group. This currently only supports at
most two column names (i.e. two-dimensions).}

\item{xlab}{x-axis label.}

\item{ylab}{y-axis label.}

\item{data.frame}{output from \code{get_cox_res}.}
}
\value{
Forest plot of cox regression results in the ggplot framework.
}
\description{
\code{plot_cox_res} takes the output from \code{get_cox_res} and generates
a forest plot showing the hazard ratio and confidence interval of the cox
cox regression.
}
\examples{
library("survival")

in.df <- colon
endpoint <- "time"
endpoint.code <- "status"

# Run and Plot Multivariate Cox Regression on Entire data.frame
features <- c("age", "obstruct")
cox.res.df <- get_cox_res(colon, endpoint, endpoint.code, features)
plot_cox_res(cox.res.df)

# Run and Plot Multivariate Cox Regression For Each rx Group
group <- "rx"
cox.res.df <- get_cox_res(colon, endpoint, endpoint.code, features, group)
plot_cox_res(cox.res.df, group)

# Change x and y labels
plot_cox_res(cox.res.df, group, xlab = "Hazard Ratio", y = "Feature")
}

