% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_KM_curve.R
\name{plot_KM_curve}
\alias{plot_KM_curve}
\title{Create a Kaplan-Meier plot using ggplot2}
\usage{
plot_KM_curve(sfit, returns = FALSE, xlabs = "Time",
  ylabs = "survival probability", ystratalabs = NULL, ystrataname = NULL,
  timeby = 100, main = "Kaplan-Meier Plot", pval = TRUE, ...)
}
\arguments{
\item{sfit}{a \code{\link[survival]{survfit}} object}

\item{returns}{logical: if \code{TRUE}, return an ggplot object}

\item{xlabs}{x-axis label}

\item{ylabs}{y-axis label}

\item{ystratalabs}{The strata labels. \code{Default = levels(summary(sfit)$strata)}}

\item{ystrataname}{The legend name. Default = “Strata”}

\item{timeby}{numeric: control the granularity along the time-axis}

\item{main}{plot title}

\item{pval}{logical: add the pvalue to the plot?}

\item{pval}{logical: add the pvalue to the plot?}
}
\value{
a ggplot is made. if returns=TRUE, then an ggplot object
  is returned
}
\description{
This a modified version of the original ggkm code by Ahijit Dasgupta. It
generates a Kaplan-Meier plot using ggplot2
}
\details{
This function replicates the functionality provided in ggfortify. Strongly 
suggest users to use that. Additionally, \code{plot_KM_curve} currently does 
not work when the input survival object contains only one group.
}
\examples{
fit <- survival::survfit(survival::Surv(time,status) ~ rx, data = survival::colon)
plot_KM_curve(fit, timeby = 500)
}
\author{
Abhijit Dasgupta \url{https://gist.github.com/araastat/9927677}
}

