% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_auc.R
\name{PlotAUC}
\alias{PlotAUC}
\title{Plot Area Under the Kaplan-Meier Curve.}
\usage{
PlotAUC(
  data,
  arm_label = "Placebo",
  arm_name = "arm",
  color = "#C65842",
  legend_pos = "top",
  return_surv = TRUE,
  status_name = "status",
  tau = NULL,
  time_name = "time",
  title = NULL,
  which_arm = 0,
  x_breaks = NULL,
  x_labs = NULL,
  x_name = "Time",
  x_max = NULL,
  y_name = "Survival",
  y_lim = c(0, 1)
)
}
\arguments{
\item{data}{Data including time, status, arm.}

\item{arm_label}{Label for the arm.}

\item{arm_name}{Name of arm column.}

\item{color}{Color.}

\item{legend_pos}{Legend position.}

\item{return_surv}{Logical, TRUE for survival, FALSE for cumulative incidence.}

\item{status_name}{Name of status column.}

\item{tau}{Truncation time for shading.}

\item{time_name}{Name of time column.}

\item{title}{Plot title.}

\item{which_arm}{0 for control, 1 for treatment.}

\item{x_breaks}{X-axis breaks.}

\item{x_labs}{X-axis labels.}

\item{x_name}{X-axis name.}

\item{x_max}{X-axis upper limit; may differ from tau.}

\item{y_name}{Y-axis name.}

\item{y_lim}{Y-axis limits.}
}
\value{
ggplot.
}
\description{
Plot Area Under the Kaplan-Meier Curve.
}
