% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nars.R
\name{PlotNARs}
\alias{PlotNARs}
\title{Plot Numbers at Risk}
\usage{
PlotNARs(
  data,
  x_breaks,
  arm_name = "arm",
  status_name = "status",
  time_name = "time",
  x_labs = NULL,
  x_max = NULL,
  x_name = NULL,
  y_labs = c("Ctrl", "Trt")
)
}
\arguments{
\item{data}{Data.frame.}

\item{x_breaks}{X-axis breaks.}

\item{arm_name}{Name of arm column.}

\item{status_name}{Name of status column.}

\item{time_name}{Name of time column.}

\item{x_labs}{X-axis labels.}

\item{x_max}{X-axis upper limit.}

\item{x_name}{X-axis name.}

\item{y_labs}{Y-axis labels.}
}
\value{
ggplot.
}
\description{
Plot Numbers at Risk
}
