% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cum_haz.R
\name{PlotTwoSampleCH}
\alias{PlotTwoSampleCH}
\title{Plot Two Sample Cumulative Hazard}
\usage{
PlotTwoSampleCH(
  data,
  arm_name = "arm",
  color_labs = c("Ctrl", "Trt"),
  color_ctrl = "#EFC000FF",
  color_trt = "#6385B8",
  plot_haz_haz = TRUE,
  status_name = "status",
  tau = NULL,
  time_name = "time",
  title = NULL,
  x_breaks = NULL,
  x_labs = NULL,
  x_lim = NULL,
  x_name = NULL,
  x_max = NULL,
  y_name = NULL,
  y_lim = NULL
)
}
\arguments{
\item{data}{Data.frame containing time, status, and arm.}

\item{arm_name}{Name of arm column.}

\item{color_labs}{Color labels.}

\item{color_ctrl}{Color for control arm.}

\item{color_trt}{Color for treatment arm.}

\item{plot_haz_haz}{Logical, plot hazard vs. hazard? If FALSE, plots hazards vs. time.}

\item{status_name}{Name of status column.}

\item{tau}{Truncation time.}

\item{time_name}{Name of time column.}

\item{title}{Plot title.}

\item{x_breaks}{X-axis breaks.}

\item{x_labs}{X-axis labels.}

\item{x_lim}{X-axis limits.}

\item{x_name}{X-axis name.}

\item{x_max}{X-axis upper limit.}

\item{y_name}{Y-axis name.}

\item{y_lim}{Y-axis limits.}
}
\value{
ggplot.
}
\description{
Plot Two Sample Cumulative Hazard
}
