% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{print.OneSample}
\alias{print.OneSample}
\title{Print Method for OneSample Object.}
\usage{
\method{print}{OneSample}(x, ...)
}
\arguments{
\item{x}{An object of class \code{OneSample}.}

\item{...}{Unused.}
}
\description{
Print method for objects of class \code{OneSample}.
}
