% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{print.OneSampleCIC}
\alias{print.OneSampleCIC}
\title{Print Method for OneSampleCIC Object}
\usage{
\method{print}{OneSampleCIC}(x, ...)
}
\arguments{
\item{x}{An object of class \code{OneSampleCIC}.}

\item{...}{Unused.}
}
\description{
Print method for objects of class \code{OneSampleCIC}.
}
