% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{print.OneSampleSurv}
\alias{print.OneSampleSurv}
\title{Print Method for OneSampleSurv Object}
\usage{
\method{print}{OneSampleSurv}(x, ...)
}
\arguments{
\item{x}{An object of class \code{OneSampleSurv}.}

\item{...}{Unused.}
}
\description{
Print method for objects of class \code{OneSampleSurv}.
}
