% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{print.SurvSummary}
\alias{print.SurvSummary}
\title{Print Method for SurvSummary Object.}
\usage{
\method{print}{SurvSummary}(x, ...)
}
\arguments{
\item{x}{An object of class \code{SurvSummary}.}

\item{...}{Unused.}
}
\description{
Print method for objects of class \code{SurvSummary}.
}
