% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{print.TwoSample}
\alias{print.TwoSample}
\title{Print Method for TwoSample Object.}
\usage{
\method{print}{TwoSample}(x, ...)
}
\arguments{
\item{x}{An object of class \code{TwoSample}.}

\item{...}{Unused.}
}
\description{
Print method for objects of class \code{TwoSample}.
}
