% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_rates.R
\name{RateRatio}
\alias{RateRatio}
\title{Calculate Rate Ratio}
\usage{
RateRatio(
  rates,
  alpha = 0.05,
  arm_name = "arm",
  rate_name = "rate",
  se_name = "se"
)
}
\arguments{
\item{rates}{Data.frame.}

\item{alpha}{Type I error.}

\item{arm_name}{Name of arm column.}

\item{rate_name}{Name of rate column.}

\item{se_name}{Name of standard error column.}
}
\value{
Data.frame with one row: \code{stat} = "rr", \code{est}, \code{se}, \code{lower}, \code{upper}, \code{p}.
}
\description{
Calculate the rate ratio comparing two groups. Input data.frame should
contain arm (0 and 1), event rate, and standard error.
}
