% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{RMST}
\alias{RMST}
\title{Calculate Restricted Mean Survival Time}
\usage{
RMST(status, time, extend = FALSE, tau = NULL)
}
\arguments{
\item{status}{Status, coded as 0 for censoring, 1 for death.}

\item{time}{Observation time.}

\item{extend}{Extend AUC calculation if tau exceeds max(time)?}

\item{tau}{Truncation time.}
}
\value{
Numeric restricted mean survival time.
}
\description{
Calculate Restricted Mean Survival Time
}
