% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{SurvSummary-class}
\alias{SurvSummary-class}
\title{Survival Summary Object.}
\description{
Survival Summary Object.
}
\section{Slots}{

\describe{
\item{\code{Marginal}}{Per arm statistics.}

\item{\code{Contrasts}}{Contrasts of the marginal statistics.}
}}

