% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurvUtils-package.R
\docType{package}
\name{SurvUtils-package}
\alias{SurvUtils}
\alias{SurvUtils-package}
\title{SurvUtils: Utility Functions for Survival Analysis}
\description{
Utility functions for survival analysis.
}
\author{
\strong{Maintainer}: Zachary McCaw \email{zmccaw@alumni.harvard.edu} (\href{https://orcid.org/0000-0002-2006-9828}{ORCID})

}
\keyword{internal}
