% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survutils.R
\name{survutils}
\alias{survutils}
\title{Predict on survival data}
\usage{
survutils(filename, type, rhor = 1, rhoc = 1, qr = 2, qc = 2,
  tol = 0.001, ntrees = 1000, nfolds = 5, ntimes = 0,
  optimizer = "GLMNET", nfeatures = 0)
}
\arguments{
\item{filename}{Path to survival data file. Data should have the features in the format (time, status and other feature columns)}

\item{type}{Type of function to be applied. Values are {"RSF", "SUPERPC", "BOOSTCI", "SCORE", "TREC", "REC"}}

\item{rhor}{Row regularization parameter. Default value is 1. Used only in type "TREC"}

\item{rhoc}{Column regularization parameter. Default value is 1. Used only in type "TREC"}

\item{qr}{Norm on the row. When 1 it is L_1 norm and when 2 it is L_2 norm. This value should be kept <=2 for reducing computational cost. Used only in type "TREC"}

\item{qc}{Norm on the column. When 1 it is L_1 norm and when 2 it is L_2 norm. This value should be kept <=2 for reducing computational cost. Used only in type "TREC"}

\item{tol}{Convergence tolerance. Default value is 1e-3. Used only in type "TREC"}

\item{ntrees}{Number of trees for BOOSTCI model. Default is 1000}

\item{nfolds}{Number of folds for BOOSTCI cross validation. Default is 5}

\item{ntimes}{Number of distinct times of events to be considered for SCORE. Computation gets slow with large values. If no value is specified all times are taken}

\item{optimizer}{Type of optimization to apply for SCORE. Values are {"GLMNET", "MRCE"}. Default is "GLMNET". "MRCE" cannot run for more than 1000 features.}

\item{nfeatures}{Number of features for SCORE. If no value is specified all features are taken}
}
\description{
Predict on survival data
}

