% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{TwoSample-class}
\alias{TwoSample-class}
\title{Two Sample Survival Summary Object.}
\description{
Two Sample Survival Summary Object.
}
\section{Slots}{

\describe{
\item{\code{Marginal}}{Per arm statistics.}

\item{\code{Contrasts}}{Contrasts of the marginal statistics.}
}}

