% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{show,TwoSample-method}
\alias{show,TwoSample-method}
\title{Show Method for TwoSample Object}
\usage{
\S4method{show}{TwoSample}(object)
}
\arguments{
\item{object}{An object of class \code{TwoSample}.}
}
\description{
Show Method for TwoSample Object
}
