% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cstat.R
\name{WeightedCstat}
\alias{WeightedCstat}
\title{Calculate Weighted Concordance}
\usage{
WeightedCstat(
  train_data,
  test_data,
  risk_name = "risk",
  status_name = "status",
  tau = NULL,
  time_name = "time"
)
}
\arguments{
\item{train_data}{Data for estimating the censoring distribution. Must
include \code{status} and \code{time}.}

\item{test_data}{Data for evaluating the risk score. Must include \code{risk},
\code{status}, and \code{time}.}

\item{risk_name}{Name of the risk column.}

\item{status_name}{Name of status column.}

\item{tau}{Truncation time; default is the 95th percentile of \code{train_data$time}.}

\item{time_name}{Name of time column.}
}
\value{
Numeric weighted C-statistic (between 0 and 1).
}
\description{
C-statistic in the presence of censoring using inverse probability of
censoring weighting. See Uno et al. (2011),
\url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC3079915/}.
}
