% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CalcCIC}
\alias{CalcCIC}
\title{Calculate CIC}
\usage{
CalcCIC(status, time)
}
\arguments{
\item{status}{Status, coded as 0 for censoring, 1 for an event, 2 for death.}

\item{time}{Observation time.}
}
\value{
Tabulate cumulative incidence curve.
}
\description{
Estimate the cumulative incidence curve. Specifically:
\eqn{F_{1}(t) = P(T \leq t, \delta = 1).}
}
