% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazard_ratio.R
\name{CalcHR}
\alias{CalcHR}
\title{Calculate Hazard Ratio}
\usage{
CalcHR(data, arm_name = "arm", status_name = "status", time_name = "time")
}
\arguments{
\item{data}{Reconstructed data.}

\item{arm_name}{Name of arm column.}

\item{status_name}{Name of status column.}

\item{time_name}{Name of time column.}
}
\value{
Data.frame.
}
\description{
Calculate the hazard ratio, comparing two treatment arms, its confidence
interval and p-value. Also provides the Schoenfeld residual test of the
proportional hazards assumption.
}
