% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumulative_incidence.R
\name{CICInfluence}
\alias{CICInfluence}
\title{Cumulative Incidence Influence}
\usage{
CICInfluence(data, tau = NULL, status_name = "status", time_name = "time")
}
\arguments{
\item{data}{Data.frame.}

\item{tau}{Truncation time at which to calculate the influence.}

\item{status_name}{Name of status column.}

\item{time_name}{Name of time column.}
}
\value{
Data.frame with an additional column, \code{influence}.
}
\description{
Calculate cumulative incidence influence function for each observation.
}
