% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumulative_incidence.R
\name{CICurves}
\alias{CICurves}
\title{Generate Cumulative Incidence Curves}
\usage{
CICurves(data, alpha = 0.05, status_name = "status", time_name = "time")
}
\arguments{
\item{data}{Data.frame.}

\item{alpha}{Type I error.}

\item{status_name}{Name of status column.}

\item{time_name}{Name of time column.}
}
\value{
OneSampleSurv object.
}
\description{
Intended for data from a single sample. Status should be coded as 0 for
censoring, 1 for the event of interest, and 2 for the competing risk (e.g.
death).
}
