% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_rmst.R
\name{CompareRMSTs}
\alias{CompareRMSTs}
\title{Compare RMSTs}
\usage{
CompareRMSTs(
  data,
  alpha = 0.05,
  arm_name = "arm",
  status_name = "status",
  tau = NULL,
  time_name = "time"
)
}
\arguments{
\item{data}{Data.frame.}

\item{alpha}{Type I error.}

\item{arm_name}{Name of arm column.}

\item{status_name}{Name of status column.}

\item{tau}{Truncation time.}

\item{time_name}{Name of time column.}
}
\value{
Data.frame.
}
\description{
Compare the restricted mean survival times of two treatment arms.
}
