% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cstat.R
\name{Cstat}
\alias{Cstat}
\title{Calculate Unweighted Concordance}
\usage{
Cstat(
  data,
  pseudo_counts = TRUE,
  risk_name = "risk",
  status_name = "status",
  tau = NULL,
  time_name = "time"
)
}
\arguments{
\item{data}{Data frame including risk, status, time.}

\item{pseudo_counts}{Include pseudo counts?}

\item{risk_name}{Name of the risk column.}

\item{status_name}{Name of status column.}

\item{tau}{Truncation time.}

\item{time_name}{Name of time column.}
}
\value{
Data.frame.
}
\description{
Calculate Unweighted Concordance
}
