% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cstat.R
\name{Cstat}
\alias{Cstat}
\title{Calculate Concordance}
\usage{
Cstat(
  train_data,
  test_data,
  risk_name = "risk",
  status_name = "status",
  tau = NULL,
  time_name = "time"
)
}
\arguments{
\item{train_data}{Data for estimating the censoring distribution. Should
include status and time.}

\item{test_data}{Data for evaluating the risk score. Should include risk,
stats, and time.}

\item{risk_name}{Name of the risk column.}

\item{status_name}{Name of status column.}

\item{tau}{Truncation time.}

\item{time_name}{Name of time column.}
}
\value{
Data.frame.
}
\description{
Calculate the C-statistic in the presence of censoring using the method of:
\url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC3079915/}.
}
