% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_data.R
\name{GenCRData}
\alias{GenCRData}
\title{Generate Competing Risks Data}
\usage{
GenCRData(
  base_death_rate = 0.25,
  beta_death = NULL,
  base_event_rate = 1,
  beta_event = NULL,
  censoring_rate = 0.25,
  covariates = NULL,
  frailty_variance = 0,
  min_death_rate = 0,
  min_event_rate = 0.05,
  n = 100,
  simple = TRUE,
  tau = 4
)
}
\arguments{
\item{base_death_rate}{Baseline arrival for for death (the competing risk).}

\item{beta_death}{Numeric vector of log rate ratios for the death rate.}

\item{base_event_rate}{Baseline arrival rate for events.}

\item{beta_event}{Numeric vector of log rate ratios for the event rate.}

\item{censoring_rate}{Arrival rate for the censoring time.}

\item{covariates}{Numeric design matrix.}

\item{frailty_variance}{Variance of the gamma frailty.}

\item{min_death_rate}{Minimum subject-specific event rate. Must be non-negative.}

\item{min_event_rate}{Minimum subject-specific event rate. Must be positive.}

\item{n}{Number of subjects. Overwritten by \code{nrow(covariates)} if covariates are provided.}

\item{simple}{Return only the index, time, and status? If FALSE, returns additional data.}

\item{tau}{Truncation time.}
}
\value{
Data.frame.
}
\description{
Generate Competing Risks Data
}
