% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_data.R
\name{GenData}
\alias{GenData}
\title{Purpose: Simulate Data.
Updated: 2022-03-19
Generate Data}
\usage{
GenData(
  base_event_rate = 1,
  beta_event = NULL,
  censoring_rate = 0.25,
  covariates = NULL,
  frailty_variance = 0,
  min_event_rate = 0.05,
  n = 100,
  tau = 4
)
}
\arguments{
\item{base_event_rate}{Baseline arrival rate for events.}

\item{beta_event}{Numeric vector of log rate ratios for the event rate.}

\item{censoring_rate}{Arrival rate for the censoring time.}

\item{covariates}{Numeric design matrix.}

\item{frailty_variance}{Variance of the gamma frailty.}

\item{min_event_rate}{Minimum subject-specific event rate. Must be positive.}

\item{n}{Number of subjects. Overwritten by \code{nrow(covariates)} if covariates are provided.}

\item{tau}{Truncation time.}
}
\value{
Data.frame.
}
\description{
Purpose: Simulate Data.
Updated: 2022-03-19
Generate Data
}
