% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_pseudo_values.R
\name{GenPseudo}
\alias{GenPseudo}
\title{Generate Pseudo-values}
\usage{
GenPseudo(
  data,
  tau = NULL,
  type = "prob",
  status_name = "status",
  time_name = "time"
)
}
\arguments{
\item{data}{Data.frame.}

\item{tau}{Truncation time.}

\item{type}{Type of pseudo-value, select "prob" for probability and "rmst"
for restricted mean survival time.}

\item{status_name}{Name of status column.}

\item{time_name}{Name of time column.}
}
\value{
Data.frame including the influence function and pseudo-value for each
observation.
}
\description{
Generates pseudo-values for each patient as the target parameter plus the
influence function.
}
