% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_rates.R
\name{OddsRatio}
\alias{OddsRatio}
\title{Calculate Rate Odds Ratio}
\usage{
OddsRatio(
  rates,
  alpha = 0.05,
  arm_name = "arm",
  rate_name = "rate",
  se_name = "se"
)
}
\arguments{
\item{rates}{Data.frame.}

\item{alpha}{Type I error.}

\item{arm_name}{Name of arm column.}

\item{rate_name}{Name of rate column.}

\item{se_name}{Name of standard error column.}
}
\value{
Data.frame
}
\description{
Calculate the odds ratio comparing two treatment arms.
}
