% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_sample.R
\name{OneSampleCIC}
\alias{OneSampleCIC}
\title{One Sample Cumulative Incidence Curve}
\usage{
OneSampleCIC(
  data,
  tau = NULL,
  alpha = 0.05,
  status_name = "status",
  time_name = "time"
)
}
\arguments{
\item{data}{Data.frame.}

\item{tau}{Truncation time.}

\item{alpha}{Type I error.}

\item{status_name}{Name of status column.}

\item{time_name}{Name of time column.}
}
\value{
Data.frame.
}
\description{
For calculating cumulative incidence in the presence in the presence of a
competing risk, status is assumed to have values of 0 for censoring, 1 for
the event of interest, and 2 for the competing risk.
}
