% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_sample.R
\name{OneSamplePercentiles}
\alias{OneSamplePercentiles}
\title{One Sample Percentiles}
\usage{
OneSamplePercentiles(
  data,
  p = 0.5,
  alpha = 0.05,
  status_name = "status",
  time_name = "time"
)
}
\arguments{
\item{data}{Data.frame.}

\item{p}{Percentile.}

\item{alpha}{Type I error.}

\item{status_name}{Name of status column.}

\item{time_name}{Name of time column.}
}
\value{
Data.frame.
}
\description{
One Sample Percentiles
}
