% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cum_haz.R
\name{PlotOneSampleCH}
\alias{PlotOneSampleCH}
\title{Plot One Sample Cumulative Hazard}
\usage{
PlotOneSampleCH(
  data,
  color = "#0F9D58",
  status_name = "status",
  tau = NULL,
  time_name = "time",
  title = NULL,
  x_breaks = NULL,
  x_labs = NULL,
  x_name = "Time",
  x_max = NULL,
  y_name = "Cumulactive Hazard",
  y_lim = NULL
)
}
\arguments{
\item{data}{Data.frame containing time and status.}

\item{color}{Color for KM curve.}

\item{status_name}{Name of status column.}

\item{tau}{Truncation time.}

\item{time_name}{Name of time column.}

\item{title}{Plot title.}

\item{x_breaks}{X-axis breaks.}

\item{x_labs}{X-axis labels.}

\item{x_name}{X-axis name.}

\item{x_max}{X-axis upper limit.}

\item{y_name}{Y-axis name.}

\item{y_lim}{Y-axis limits.}
}
\value{
ggplot.
}
\description{
Plot One Sample Cumulative Hazard
}
