% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_kaplan_meier.R
\name{PlotOneSampleKM}
\alias{PlotOneSampleKM}
\title{Plot Survival Data.}
\usage{
PlotOneSampleKM(
  data,
  color = "#0F9D58",
  fit = NULL,
  plot_surv = TRUE,
  status_name = "status",
  tau = NULL,
  time_name = "time",
  title = NULL,
  x_breaks = NULL,
  x_labs = NULL,
  x_name = "Time",
  x_max = NULL,
  y_name = "Survival",
  y_lim = c(0, 1)
)
}
\arguments{
\item{data}{Data.frame containing time and status.}

\item{color}{Color for KM curve.}

\item{fit}{Parametric fit from the Temporal package.}

\item{plot_surv}{Logical, TRUE for survival curves, FALSE for cumulative incidence.}

\item{status_name}{Name of status column.}

\item{tau}{Truncation time.}

\item{time_name}{Name of time column.}

\item{title}{Plot title.}

\item{x_breaks}{X-axis breaks.}

\item{x_labs}{X-axis labels.}

\item{x_name}{X-axis name.}

\item{x_max}{X-axis upper limit.}

\item{y_name}{Y-axis name.}

\item{y_lim}{Y-axis limits.}
}
\value{
ggplot.
}
\description{
Plot Survival Data.
}
