% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_nars.R
\name{PlotOneSampleNARs}
\alias{PlotOneSampleNARs}
\title{Plot Two Sample Numbers at Risk}
\usage{
PlotOneSampleNARs(
  data,
  specified_nars = NULL,
  status_name = "status",
  time_name = "time",
  x_breaks = NULL,
  x_labs = NULL,
  x_max = NULL,
  x_name = NULL,
  y_lab = "NAR"
)
}
\arguments{
\item{data}{Data.frame.}

\item{specified_nars}{Optional data.frame of specified NARs. Should include
"time" and "nar". Overrides \code{data} if present.}

\item{status_name}{Name of status column.}

\item{time_name}{Name of time column.}

\item{x_breaks}{X-axis breaks.}

\item{x_labs}{X-axis labels.}

\item{x_max}{X-axis upper limit.}

\item{x_name}{X-axis name.}

\item{y_lab}{Y-axis label.}
}
\value{
ggplot.
}
\description{
Plot Two Sample Numbers at Risk
}
