% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmst.R
\name{RMSTInfluence}
\alias{RMSTInfluence}
\title{RMST Influence}
\usage{
RMSTInfluence(data, tau = NULL, status_name = "status", time_name = "time")
}
\arguments{
\item{data}{Data.frame.}

\item{tau}{Truncation time at which to calculate the influence.}

\item{status_name}{Name of status column.}

\item{time_name}{Name of time column.}
}
\value{
Numeric vector.
}
\description{
Calculate RMST influence function for each observation.
}
