% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumulative_incidence.R
\name{TabulateCIC}
\alias{TabulateCIC}
\title{Purpose: Cumulative incidence estimator.
Updated: 2024-08-04
Tabulate Cumulative Incidence}
\usage{
TabulateCIC(data, status_name = "status", time_name = "time", alpha = 0.05)
}
\arguments{
\item{data}{Data.frame.}

\item{status_name}{Name of status column.}

\item{time_name}{Name of time column.}

\item{alpha}{Type I error, for confidence intervals.}
}
\value{
Data.frame.
}
\description{
Status should be coded as 0 for censoring, 1 for the event of interest,
and 2 for the competing risk (e.g. death).
}
