% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kaplan_meier.R
\name{TabulateKM}
\alias{TabulateKM}
\title{Purpose: Kaplan-Meier estimator.
Updated: 2023-09-26
Tabulate Kaplan-Meier}
\usage{
TabulateKM(data, status_name = "status", time_name = "time", alpha = 0.05)
}
\arguments{
\item{data}{Data.frame.}

\item{status_name}{Name of status column.}

\item{time_name}{Name of time column.}

\item{alpha}{Type I error, for confidence intervals.}
}
\value{
Data.frame.
}
\description{
Purpose: Kaplan-Meier estimator.
Updated: 2023-09-26
Tabulate Kaplan-Meier
}
